pyproject.toml
[project]
name = "lab8"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "mysql-connector-python",
]
